+++ src/views/_shared/Header.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import LaptopMacIcon from "@material-ui/icons/LaptopMac";
 import UserMenu from "./UserMenu";
 import { NavLink } from "react-router-dom";
-import { ThemeContext } from "../context/contexts";
+import { ThemeContext } from "../../context/contexts";
 
 const Header = (props: any) => {
    const [darkMode, setDarkMode] = useState(false);
